test: split isHexColor validator tests
